5: small refactoring (finally)
